Correct terraform destroy job, add groups
TL;DR
=====
- Correct the `terraform-destroy` job
- Group jobs in `create` or `destroy` headings

Detail
======
`terraform-destroy` task now executes correctly.

Groupings of jobs were created to make the `create` versus `destroy`
operations easier to identify. This makes it a conscious effort to move
to the `destory` group to select `terraform-destroy`.

NOTE
====
The terraform will need to be corrected to not delete the resource
group, as terraform tries to persist the `tfstate` back to the Azure
Blobstore container that it's configured to use.